delete UseHook-directory, refactor UserAccount-directory(in Components, create ModalComponents-directory && CardComponents-directory)
