+++ .vscode/settings.json
@@ -12,6 +12,7 @@
         "jewellery",
         "junghan",
         "Junghans",
+        "MOMO",
         "womens"
     ]
 }
