+++ src/Services/AccountServices.js
@@ -1,4 +1,4 @@
-import { get, ref, onValue, remove } from "firebase/database";
+import { get, ref } from "firebase/database";
 import { database } from "../FirebaseConfigs/Firesbase"; // Assuming you have initialized your Firebase app and exported the database instance as 'database'
 
 export const DeliveryServices = async (userId) => {
